Improve responsive design for small windows
The game board will now be correctly centered in the screen and scaled
to the maximum size available, at all screen sizes.

The font size will be reduced at small screen sizes, and other interface
elements (such as the indicator of which player is to move) scale in
proportion to the font size.