Update to latest version of Vue-Webpack
While building for release, I encountered an error with Babel. This
prompted an npm update. This broke the project. This led to replacing
the the Vue Webpack template with the latest version. This did not fix
the project, and introduced some new build warnings. Hooray for
progress.

The build warnings are fixed. The problem that was preventing rendering
- use of the spread operator instead of Array.from() - has also been
fixed.