Add web font definition of title font
Previously I had mistakenly remembered using Linux Libertine Regular to
the title font. I later realised when testing on other devices that it
was using Linux Libertine Display, and I didn't notice because it was
using a local copy instead of a webfont. Since no webfont copy of the
Display version was defined, it would fall back to a standard local
serif font if it couldn't find a local copy. It looked bad.

Linux Libertine Display has been made available on
canmom.github.io/fonts to enable this change.